+++ Src/Private/Add-DiaTShapeLine.ps1
@@ -1,7 +1,13 @@
 function Add-DiaTShapeLine {
     <#
     .SYNOPSIS
-        Function to create a T shape object in the diagram.
+        Function to create an T-shaped ( T ) object in the diagram, connecting specified nodes with customizable line styles and attributes.
+            Example:
+                                (TShapeMiddleUpper)
+               (TShapeStart)o___o___o(TShapeEnd)
+                                |
+                                o
+                                (TShapeMiddleDown)
     .DESCRIPTION
         Function to create a T shape object in the diagram.
     .NOTES
@@ -23,35 +29,41 @@ function Add-DiaTShapeLine {
             Mandatory = $false,
             HelpMessage = 'Please provide a string to be used as Start Node Name'
         )]
-        [string] $StartName = 'TStart',
+        [string] $StartName = 'TShapeStart',
 
         [Parameter(
             Mandatory = $false,
             HelpMessage = 'Please provide a string to be used as End Node Name'
         )]
-        [string] $EndName = 'TEnd',
+        [string] $EndName = 'TShapeEnd',
 
         [Parameter(
             Mandatory = $false,
             HelpMessage = 'Please provide a string to be used as End Node Name'
         )]
-        [string] $MiddleTop = 'TMiddleTop',
+        [string] $MiddleTop = 'TShapeMiddleUpper',
 
         [Parameter(
             Mandatory = $false,
             HelpMessage = 'Please provide a string to be used as End Node Name'
         )]
-        [string] $MiddleDown = 'TMiddleDown',
+        [string] $MiddleDown = 'TShapeMiddleDown',
 
         [Parameter(
             Mandatory = $false,
             HelpMessage = 'Please provide a string to be used as End Node Name'
+        )]
+        [ValidateSet(
+            'none', 'normal', 'inv', 'dot', 'invdot', 'odot', 'invodot', 'diamond', 'odiamond', 'ediamond', 'crow', 'box', 'obox', 'open', 'halfopen', 'empty', 'invempty', 'tee', 'vee', 'icurve', 'lcurve', 'rcurve', 'icurve', 'box', 'obox', 'diamond', 'odiamond', 'ediamond', 'crow', 'tee', 'vee', 'dot', 'odot', 'inv', 'invodot', 'invempty', 'invbox', 'invodiamond', 'invtee', 'invvee', 'none'
         )]
         [string] $Arrowtail = 'none',
 
         [Parameter(
             Mandatory = $false,
             HelpMessage = 'Please provide a string to be used as End Node Name'
+        )]
+        [ValidateSet(
+            'none', 'normal', 'inv', 'dot', 'invdot', 'odot', 'invodot', 'diamond', 'odiamond', 'ediamond', 'crow', 'box', 'obox', 'open', 'halfopen', 'empty', 'invempty', 'tee', 'vee', 'icurve', 'lcurve', 'rcurve', 'icurve', 'box', 'obox', 'diamond', 'odiamond', 'ediamond', 'crow', 'tee', 'vee', 'dot', 'odot', 'inv', 'invodot', 'invempty', 'invbox', 'invodiamond', 'invtee', 'invvee', 'none'
         )]
         [string] $Arrowhead = 'none',
 
@@ -67,13 +79,26 @@ function Add-DiaTShapeLine {
             HelpMessage = 'Please provide a string to be used as End Node Name'
         )]
         [ValidateRange(1, 10)]
-        [int] $LineSize = 1,
+        [int] $LineLength = 1,
+
+        [Parameter(
+            Mandatory = $false,
+            HelpMessage = 'Width of the line (penwidth), from 1 to 10.'
+        )]
+        [ValidateRange(1, 10)]
+        [int] $LineWidth = 1,
 
         [Parameter(
             Mandatory = $false,
             HelpMessage = 'Please provide a color to be used in the line. Default is black. Supported color https://graphviz.org/doc/info/colors.html'
         )]
-        [string] $LineColor = "black"
+        [string] $LineColor = "black",
+
+        [Parameter(
+            Mandatory = $false,
+            HelpMessage = 'Enables debug mode for icons, highlighting the table in red.'
+        )]
+        [bool] $IconDebug = $false
     )
 
     begin {
@@ -81,12 +106,29 @@ function Add-DiaTShapeLine {
 
     process {
         try {
-            Node $StartName, $MiddleDown, $EndName @{shape = 'none'; fixedsize = 'true'; width = .001 ; height = .001; fillColor = 'transparent'; style = 'invis' }
-            Node $MiddleTop @{shape = 'point'; fixedsize = 'true'; width = .001 ; height = .001; fillColor = 'transparent'; style = 'filled' }
+            if ($IconDebug) {
+                $Shape = 'plain'
+                $fillColor = 'red'
+                $Style = 'filled'
+                $Color = 'black'
+                $LineColor = 'red'
+            } else {
+                $Shape = 'point'
+                $fillColor = 'transparent'
+                $Style = 'invis'
+                $Color = $LineColor
+            }
+            if ($IconDebug) {
+                Node $StartName, $MiddleDown, $EndName, $MiddleTop @{color = $Color; shape = $Shape; fillColor = $fillColor; style = $Style }
+            } else {
+                Node $StartName, $MiddleDown, $EndName @{color = $Color; shape = $Shape; fixedsize = 'true'; width = .001 ; height = .001; fillColor = $fillColor; style = $Style }
+                Node $MiddleTop @{color = $Color; shape = $Shape; fixedsize = 'true'; width = .001 ; height = .001; fillColor = $fillColor; style = $Style }
+            }
+
             Rank $StartName, $MiddleTop, $EndName
-            Edge -From $StartName -To $MiddleTop @{minlen = $LineSize; arrowtail = $Arrowtail; arrowhead = $Arrowhead; style = $LineStyle; color = $LineColor }
-            Edge -From $MiddleTop -To $EndName @{minlen = $LineSize; arrowtail = $Arrowtail; arrowhead = $Arrowhead; style = $LineStyle; color = $LineColor }
-            Edge -From $MiddleTop -To $MiddleDown @{minlen = $LineSize; arrowtail = $Arrowtail; arrowhead = $Arrowhead; style = $LineStyle; color = $LineColor }
+            Edge -From $StartName -To $MiddleTop @{minlen = $LineLength; arrowtail = $Arrowtail; arrowhead = $Arrowhead; style = $LineStyle; color = $LineColor; penwidth = $LineWidth }
+            Edge -From $MiddleTop -To $EndName @{minlen = $LineLength; arrowtail = $Arrowtail; arrowhead = $Arrowhead; style = $LineStyle; color = $LineColor; penwidth = $LineWidth }
+            Edge -From $MiddleTop -To $MiddleDown @{minlen = $LineLength; arrowtail = $Arrowtail; arrowhead = $Arrowhead; style = $LineStyle; color = $LineColor; penwidth = $LineWidth }
 
 
         } catch {